feat: package ExternalViewSwitcher components
