+++ src/containers/Agenda/Agenda.tsx
@@ -1,5 +1,5 @@
 import React, { FC, useState } from 'react'
-import { Paper, Button, ButtonGroup, Typography } from '@material-ui/core'
+import { Paper, Typography } from '@material-ui/core'
 import { ViewState } from '@devexpress/dx-react-scheduler'
 import {
   Scheduler,
@@ -11,28 +11,13 @@ import {
   AppointmentTooltip,
   AllDayPanel,
 } from '@devexpress/dx-react-scheduler-material-ui'
+import ExternalViewSwitcher from './components/ExternalViewSwitcher/ExternalViewSwitcher'
 import { appointments } from './mock'
-import { ViewDate } from './constants'
 import useStyles from './styles'
-
-interface IExternalViewSwitcher {
-  currentViewName: string
-  onChange: any
-}
 
 const Agenda: FC = () => {
   const [currentViewName, setCurrentViewName] = useState('Day')
   const classes = useStyles()
-
-  const ExternalViewSwitcher: FC<IExternalViewSwitcher> = ({ onChange }) => (
-    <ButtonGroup color="secondary" aria-label="outlined secondary button group">
-      {ViewDate.map(val => (
-        <Button key={val} onClick={() => onChange(val)}>
-          {val}
-        </Button>
-      ))}
-    </ButtonGroup>
-  )
 
   return (
     <Paper>
